MOW-592 Side handlers (still not perfect)
git-svn-id: http://dev.geogebra.org/svn/trunk/geogebra@63792 23ce0884-8a58-47d3-bc5c-ddf1cd5b9f9e

+++ common/src/main/java/org/geogebra/common/euclidian/draw/DrawPolygon.java
@@ -774,7 +774,7 @@ public class DrawPolygon extends Drawable implements Previewable {
 		fixCornerCoords(hitHandler);
 
 		if (!Double.isNaN(fixCornerX)) {
-			int width = (int) (point.getX() - fixCornerX);
+			double width = point.getX() - fixCornerX;
 			double[] currCoords = new double[6];
 			GPathIterator it = gp.getPathIterator(null);
 			int i = poly.getPointsLength();
@@ -791,7 +791,7 @@ public class DrawPolygon extends Drawable implements Previewable {
 		}
 
 		if (!Double.isNaN(fixCornerY)) {
-			int height = (int) (point.getY() - fixCornerY);
+			double height = point.getY() - fixCornerY;
 			double[] currCoords = new double[6];
 			GPathIterator it = gp.getPathIterator(null);
 			int i = poly.getPointsLength();
